1. event log refresh if system settings are changed
2. remove duplicate function in businessStrings

<==== Mercurial data ====>
Branch: events_db

+++ common/src/utils/common/rect_set.h
@@ -25,7 +25,7 @@ public:
     }
 
     int size() const {
-        return m_xSet.size() / 2;
+        return (int) m_xSet.size() / 2;
     }
 
     void insert(const QRect &rect) {
